Add LoadGlobalBlockFileInfo func to PersistGlobal
